AEGIS-Server-aegis-2024-sprint-05/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
attrs==20.2.0
beautifulsoup4==4.9.1
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.0.5
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.1.1
django-environ==0.4.5
django-extensions==2.2.9
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-model-utils==4.0.0
django-nested-admin==3.3.2
django-sendgrid-v5==0.8.1
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
draftjs-exporter==4.1.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
markdown2==2.3.10
oauthlib==3.1.0
pathspec==0.8.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.5
PyJWT==1.7.1
PyPDF2==1.26.0
python-http-client==3.3.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.4
regex==2020.10.28
reportlab==3.5.55
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.7
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
XlsxWriter==1.3.5
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.9.1
black==19.10b0
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coverage==5.1
defusedxml==0.7.1
dj-rest-auth==1.0.5
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.1.1
django-environ==0.4.5
django-extensions==2.2.9
django-markdown-deux==1.0.5
django-model-utils==4.0.0
django-sendgrid-v5==0.8.1
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
draftjs-exporter==4.1.0
drf-api-logger==1.1.14
drf-yasg==1.21.7
future==0.18.3
html5lib==1.1
idna==3.4
inflection==0.5.1
lxml==4.6.1
markdown2==2.4.10
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
psycopg2==2.8.6
PyJWT==1.7.1
PyPDF2==3.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
reportlab==3.6.6
requests==2.31.0
requests-oauthlib==1.3.1
sendgrid==6.10.0
shortuuid==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
toml==0.10.2
typed-ast==1.5.5
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
XlsxWriter==1.3.5
